swiched from files to resources #108
git-svn-id: https://svn.kwarc.info/repos/MMT/src@3700 09069c07-60fa-4895-8316-727242e252e4

+++ jEdit-mmt/src/info/kwarc/mmt/jeditsetup/Setup.scala
@@ -23,37 +23,44 @@ class Setup extends ShellExtension("jeditsetup") {
       case _ => None
     } else None
     val jeditOpt = if (l >= 2) Some(File(args(1))) else OS.jEditSettingsFolder
+    if (jeditOpt.isEmpty) {
+        println("jEdit settings folder not found")
+    }
+    if (!MMTSystem.runFromJar) {
+      println("jeditsetup can only be run from a .jar")
+      return true
+    }
     if (l <= 0 || l > 2 || installOpt.isEmpty || jeditOpt.isEmpty) {
       println(helpText)
       return true
     }
-    val programLocation = File(getClass.getProtectionDomain.getCodeSource.getLocation.getPath).getParentFile
+    val jarFile: File = MMTSystem.classFolder
+    val programLocation: File = jarFile.up
     val (deploy, fat) = if (programLocation.getName == "main") {
-      (programLocation.getParentFile, false)
+      (programLocation.up, false)
     } else {
-      (programLocation, true)
-    }
-    val setup = deploy / "jedit-plugin" / "plugin"
-    if (!setup.isDirectory) {
-      println("plugin directory not found, looked in " + setup)
-      return true
+      (jarFile, true)
     }
     val jedit = jeditOpt.get
     val install = installOpt.get
 
-    println("trying to (un)install from " + setup + " to " + jedit)
-    doIt(setup, deploy, jedit, install, fat)
+    if (install) {
+      println("trying to install to " + jedit)
+    } else
+    {
+      println("trying to uninstall from " + jedit)
+    }
+    doIt(deploy, jedit, install, fat)
     true
   }
 
   /** the actual install/uninstall process
     *
-    * @param setup   the deploy/jedit-plugin/plugin folder
     * @param jedit   the jEdit settings folder
     * @param install true/false for install/uninstall
     * @param fat     use the fat mmt.jar
     */
-  private def doIt(setup: File, deploy: File, jedit: File, install: Boolean, fat: Boolean) {
+  private def doIt(deploy: File, jedit: File, install: Boolean, fat: Boolean) {
     /** copies or deletes a file depending on install/uninstall */
     def copyFromOrDelete(dir: File, f: List[String], g: List[String]) {
       if (install) {
@@ -62,7 +69,6 @@ class Setup extends ShellExtension("jeditsetup") {
         delete(jedit / g)
       }
     }
-    def copyOrDelete(f: List[String]) = copyFromOrDelete(setup, f, f)
     // copy/delete the jars
     // see src/project/Utils.scala for sbt
     val mainJars = List("mmt-api.jar", "mmt-lf.jar", "MMTPlugin.jar", "mmt-specware.jar", "mmt-pvs.jar")
@@ -70,21 +76,38 @@ class Setup extends ShellExtension("jeditsetup") {
       "tiscaf.jar")
     val allJars = List("lfcatalog", "lfcatalog.jar") :: mainJars.map(List("main", _)) ++
       libJars.map(List("lib", _))
+    (jedit / "jars").mkdirs()
     if (fat) {
-      copyFromOrDelete(deploy, List("mmt.jar"), List("jars", "MMTPlugin.jar"))
+      copyFromOrDelete(deploy, Nil, List("jars", "MMTPlugin.jar"))
     } else {
       allJars.foreach(f => copyFromOrDelete(deploy, f, List("jars", FilePath(f).name)))
     }
     // modes
+    def copyOrDelete(f: List[String]) = {
+      val file = jedit / f
+      if (install) {
+        if (file.exists) {
+          println("warning: " + file + " already exists, skipping; you probably want to run uninstall first")
+        }
+        else {
+          println("creating " + file)
+          File.write(file, MMTSystem.getResourceAsString("/plugin/" + f.mkString("/")))
+        }
+      } else {
+        delete(file)
+      }
+    }
     // * copy/delete the mode files
-    val modeFiles = (setup / "modes").list.filter(_.endsWith(".xml"))
+    val modeFiles = List("mmt.xml", "mmtlog.xml", "msl.xml")
     modeFiles.foreach { e => copyOrDelete(List("modes", e)) }
     // * read, update, write the catalog file
-    val scat = setup / "modes" / "catalog"
+    val scat = "/plugin/modes/catalog"
     val jcat = jedit / "modes" / "catalog"
     var newCatalog: List[String] = Nil
     val modeEntries = modeFiles.map(e => "FILE=\"" + e + "\"")
     def isMMTEntry(line: String) = modeEntries.exists(line.contains)
+    def handleResourceLineWise(path : String)(proc: String => Unit) =
+      scala.io.Source.fromInputStream(MMTSystem.getResource(path)).getLines.foreach(proc)
     // read current catalog without MMT entries
     // write new catalog
     if (!jcat.exists) {
@@ -94,17 +117,19 @@ class Setup extends ShellExtension("jeditsetup") {
         if (install) {
           if (line.contains("</MODES>")) {
             // append MMT modes if installing
-            File.ReadLineWise(scat) { l => newCatalog ::= l }
+            handleResourceLineWise(scat) { l => newCatalog ::= l }
           }
         }
         if (!isMMTEntry(line))
           newCatalog ::= line
       }
     }
-    println("updating " + jcat)
-    File.WriteLineWise(jcat, newCatalog.reverse)
+    if (install || jcat.exists) {
+      println("updating " + jcat)
+      File.WriteLineWise(jcat, newCatalog.reverse)
+    }
     // abbrevs
-    val sabb = setup / "abbrevs"
+    val sabb = "/plugin/abbrevs"
     val jabb = jedit / "abbrevs"
     var newAbbrevs: List[String] = Nil
     var remove = false
@@ -123,22 +148,18 @@ class Setup extends ShellExtension("jeditsetup") {
     }
     // append MMT abbrevs if installing
     if (install) {
-      File.ReadLineWise(sabb) { l => newAbbrevs ::= l }
+      handleResourceLineWise(sabb) { l => newAbbrevs ::= l }
     }
     // write new abbrevs
-    println("updating " + jabb)
-    File.WriteLineWise(jabb, newAbbrevs.reverse)
+    if (newAbbrevs.nonEmpty) {
+      println("updating " + jabb)
+      File.WriteLineWise(jabb, newAbbrevs.reverse)
+    } else {
+      delete(jabb)
+    }
     // copy/delete pluginFolder
     val plug = List("plugins", "info.kwarc.mmt.jedit.MMTPlugin")
     copyOrDelete(plug ::: List("startup.msl"))
-    val mars = plug ::: List("mars")
-    val setupmars = setup / mars
-    if (setupmars.exists) {
-      setupmars.list.foreach { e =>
-        if (e.endsWith(".mar"))
-          copyOrDelete(mars ::: List(e))
-      }
-    }
     if (!install) {
       val d = jedit / plug
       if (d.isDirectory) {
@@ -147,15 +168,22 @@ class Setup extends ShellExtension("jeditsetup") {
       }
     }
 
+    val propsFile = jedit / "properties"
+    val propsOld = if (propsFile.exists) File.read(propsFile) else ""
+    val archKey = "info.kwarc.mmt.jedit.MMTPlugin.archives="
     if (install) controller.getOAF.map { oaf =>
       val contentFolder = oaf.root
       println("adding property for content folder " + contentFolder)
-      val propsFile = jedit / "properties"
-      val propsOld = if (propsFile.exists) File.read(propsFile) else ""
       val encoded = contentFolder.toString.replace("\\", "\\\\").replace(":", "\\:").replace("=", "\\=")
-      val newValues = "info.kwarc.mmt.jedit.MMTPlugin.archives=" + encoded
-      val propsNew = propsOld + "\n" + newValues
+      val newValues = archKey + encoded
+      val propsNew = propsOld + newValues + "\n"
       File.write(propsFile, propsNew)
+    } else {
+      println("deleting property for content folder: " + archKey)
+      val newLines = propsOld.split("\n").filter(!_.startsWith(archKey)).toList
+      if (newLines != List("")) {
+        File.WriteLineWise(propsFile, newLines)
+      } else delete(propsFile)
     }
   }
 
@@ -169,8 +197,13 @@ class Setup extends ShellExtension("jeditsetup") {
       println("warning: " + to + " already exists, skipping; you probably want to run uninstall first")
     } else {
       to.getParentFile.mkdirs
-      java.nio.file.Files.copy(from.toPath, to.toPath)
-      println("copying " + from + " to " + to)
+      if (from.exists) {
+        java.nio.file.Files.copy(from.toPath, to.toPath)
+        println("copying " + from + " to " + to)
+      } else {
+        println("warning: " + from + " does not exist! (skipping)")
+        println("jedit setup may be incomplete or inconsistent.")
+      }
     }
   }
 }
